Fix problem when removing noexistent file.

+++ slc/script/cleanup.sh
@@ -18,4 +18,6 @@ rm -f /EMPTY
 
 # Remove network rules file which will be created otherwise in the 
 # wrong way disabling ssh login completely
-rm /etc/udev/rules.d/70-persistent-net.rules
+if [ -e /etc/udev/rules.d/70-persistent-net.rules ]; then
+  rm /etc/udev/rules.d/70-persistent-net.rules
+fi
